Added scroll bar drawing in borderless mode
Also renamed and otherwise changed several function prototypes.

+++ PP2.LangtonsAnt/PP2.LangtonsAnt/grid.c
@@ -5,12 +5,12 @@
 #include "logic.h"
 #include "graphics.h"
 
-Grid *grid_new(unsigned size)
+Grid *grid_new(size_t size)
 {
 	Grid *grid = malloc(sizeof(Grid));
 	unsigned char **c = malloc(size * sizeof(unsigned char*));
 	short def_color = COLOR_WHITE;
-	unsigned i, j;
+	size_t i, j;
 	for (i = 0; i < size; ++i) {
 		c[i] = malloc(size * sizeof(unsigned char));
 		for (j = 0; j < size; ++j) {
@@ -24,7 +24,7 @@ Grid *grid_new(unsigned size)
 
 void grid_delete(Grid *grid)
 {
-	unsigned i;
+	size_t i;
 	for (i = 0; i < grid->size; ++i) {
 		free(grid->c[i]);
 	}
@@ -32,20 +32,20 @@ void grid_delete(Grid *grid)
 	free(grid);
 }
 
-static void transfer_ant(Ant *ant, unsigned old_size)
+static void transfer_ant(Ant *ant, size_t old_size)
 {
 	ant->pos.y += old_size;
 	ant->pos.x += old_size;
 }
 
-static bool is_in_old_matrix(int y, int x, unsigned old_size)
+static bool is_in_old_matrix(int y, int x, size_t old_size)
 {
 	return y >= old_size && y < 2*old_size && x >= old_size && x < 2*old_size;
 }
 
 void expand_grid(Grid *grid, Ant *ant)
 {
-	unsigned old = grid->size, size = GRID_MUL*old, i, j;
+	size_t old = grid->size, size = GRID_MUL*old, i, j;
 	unsigned char **c = malloc(size * sizeof(unsigned char*));
 	short def_color = COLOR_WHITE;
 	transfer_ant(ant, old);
